moved non click functions to download utils
